Applies a whole load of refactoring and fixes. The reader and writer types are merged with the Decoder and Encoder types respectively. This is the only place their code is ever used and it is simple enough to merge without creating a mess.
Fixes the endian code in the reader. It was assembling
words in the wrong direction. Fixes all unit tests accordingly.

Changes the API of the InstructionSet and Codecs to be little
more flexible. Also requires all instructions to implement
the Instruction interface. Its sole method is `Opcode() uint32`.
This is needed by the Encoder to figure out how to write the
instruction. This is hardly ideal and may change.

Signed-off-by: Jim Teeuwen <jimteeuwen@gmail.com>