Implements the beginnings of logical layout validation. This uses a regular expression to operate on a list of opcodes in a module's program.
While this approach works and is reasonably simple, it
has some issues. One being the use of Go's full regexp
package is overkill. We need only a small subset of its
functionality. Second, we have to treat the opcodes as a
sequence of runes. Ideally we want our own, custom regex
parser which imlpements only the bits we require and
operates directly on a list of 32-bit integers, instead
of strings.

TO BE CONTINUED.

Signed-off-by: Jim Teeuwen <jimteeuwen@gmail.com>